+++ xenon-common/src/main/java/com/vmware/xenon/common/FactoryService.java
@@ -965,6 +965,14 @@ public abstract class FactoryService extends StatelessService {
 
     @Override
     public void handleNodeGroupMaintenance(Operation maintOp) {
+
+        if (!getHost().isPeerSynchronizationEnabled()) {
+            // when synchronization is disabled, do not proceed.
+            setAvailable(true);
+            maintOp.complete();
+            return;
+        }
+
         // Reset query result limit for new synchronization cycle.
         this.selfQueryResultLimit = SELF_QUERY_RESULT_LIMIT;
         synchronizeChildServicesIfOwner(maintOp);